# doc/api.yaml
openapi: 3.0.0
info:
  title: WasaText API
  description: API semplice per la piattaforma di messaggistica WASAText
  version: 1.0.0
...
security:
  - bearerAuth: []
paths:
...
  /users/{userId}/conversations/{conversationId}:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/conversationId"
    get:
      tags:
        - conversation
      summary: Dettagli conversazione
      description: Restituisce tutti i messaggi di una conversazione.
      operationId: getConversation
      parameters:
        - $ref: "#/components/parameters/sort"
      responses:
        '200':
          description: Dettagli conversazione
          content:
            application/json:
              schema:
                type: object
                properties:
                  conversationId:
                    type: integer
                    example: 1
                  messages:
                    type: array
                    items:
                      $ref: "#/components/schemas/Message"
        '404':
          description: Conversazione non trovata
        '500':
          description: Errore server
...
  /users/{userId}/conversations/{conversationId}/messages/{messageId}/forwardMessage:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/conversationId"
      - $ref: "#/components/parameters/messageId"
    post:
      tags:
        - message
      summary: Inoltra un messaggio
      description: Inoltra un messaggio esistente a un'altra conversazione.
      operationId: forwardMessage
      requestBody:
        description: ID conversazione di destinazione
        content:
          application/json:
            schema:
              type: object
              properties:
                conversationId:
                  type: integer
                  example: 1
        required: true
      responses:
        "201":
          description: Messaggio inoltrato
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
        '400':
          description: ID conversazione o messaggio non valido
        '404':
          description: Conversazione o messaggio originale non trovato
        '500':
          description: Errore server
...
  /users/{userId}/conversations/{conversationId}/messages/{messageId}:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/conversationId"
      - $ref: "#/components/parameters/messageId"
    delete:
      tags:
        - message
      summary: Elimina un messaggio
      description: Elimina un messaggio specifico da una conversazione.
      operationId: deleteMessage
      responses:
        "204":
          description: Messaggio eliminato
        '400':
          description: ID messaggio o conversazione non valido
        '404':
          description: Messaggio o conversazione non trovati
        '500':
          description: Errore server
  /users/{userId}/groups/{groupId}/members/:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/groupId"
    put:
      tags:
        - group
      summary: Aggiungi utente a un gruppo
      description: Aggiungi un altro utente a un gruppo esistente.
      operationId: addToGroup
      requestBody:
        description: Utente da aggiungere al gruppo
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        '204':
          description: Utente aggiunto
        '400':
          description: ID gruppo o utente non valido
        '404':
          description: Gruppo o utente non trovato
        '500':
          description: Errore server

    get:
      tags:
        - group
      summary: Prendi membri del gruppo
      description: Restituisce una lista di membri di un gruppo.
      operationId: getGroupMembers
      responses:
        '200':
          description: Lista membri del gruppo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Username"
        '404':
          description: Gruppo non trovato
        '500':
          description: Errore server
  /users/{userId}/groups/{groupId}/members/me:
    parameters:
      - $ref: "#/components/parameters/userId"
      - $ref: "#/components/parameters/groupId"
    delete:
      tags:
        - group
      summary: Lascia un gruppo
      description: Permette a un utente di lasciare un gruppo.
      operationId: leaveGroup
      responses:
        "204":
          description: Utente rimosso dal gruppo
        '400':
          description: ID gruppo non valido
        '404':
          description: Gruppo non trovato
        '500':
          description: Errore server
...
  /users/{userId}/search:
    parameters:
      - $ref: "#/components/parameters/userId"
    get:
      tags:
        - user
      summary: Cerca utenti per username
      description: Cerca altri utenti per username.
      operationId: searchUsers
      parameters:
        - name: username
          in: query
          description: Username da cercare.
          required: true
          schema:
            $ref: '#/components/schemas/Username'
      responses:
        '200':
          description: Lista utenti trovati
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Username'
        '400':
          description: Query di ricerca non valida
        '404':
          description: Utente non trovato
        '500':
          description: Errore server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    Message:
      title: Message
      description: "Rappresenta un singolo messaggio."
      type: object
      properties:
        id:
          description: Identificatore unico del messaggio.
          type: integer
          example: 1
          readOnly: true
        timestamp: 
          type: string
          format: date-time
          description: Timestamp di invio del messaggio.
          example: "2023-11-05T14:30:00Z"
        content:
          description: Contenuto del messaggio.
          oneOf:
            - $ref: '#/components/schemas/Text'
            - $ref: '#/components/schemas/Photo'
        sender:
          $ref: '#/components/schemas/User'
        status:
          type: string
          enum: ["received", "read"]
          description: Stato del messaggio
          example: "read"
        commments:
          type: array
          minItems: 0
          maxItems: 500
          description: Lista di commenti sul messaggio
          items:
            $ref: '#/components/schemas/Comment'
        type:
          description: Tipo di messaggio
          type: string
          enum: ["standard", "forward", "reply"]
          example: "standard"
        replyToMessageId:
          type: integer
          description: ID del messaggio a cui si sta rispondendo.
          example: 12
    Comment:
      title: Comment
      description: "Rappresenta un singolo commento."
      type: object
      properties:
        commentId:
          description: Identificatore unico del commento.
          type: integer
          example: 1
          readOnly: true
        sender:
          $ref: '#/components/schemas/User'
        content:
          type: string
          description: Il commento come emoji
          example: "👍"
          pattern: '^.*?$'
          minLength: 1
          maxLength: 2
...
    User:
      title: User
      description: "Rappresenta un singolo utente."
      type: object
      properties:
        userId:
          $ref: '#/components/schemas/UserId'
        username:
          $ref: '#/components/schemas/Username'
    Photo:
      description: Dati immagine codificati in Base64
      type: string
      format: byte
      pattern: "^[01]+$"
      minLength: 0
      maxLength: 9999999999999999999
    Username:
      description: Username dell'utente.
      type: string
      example: "user123"
      minLength: 3
      maxLength: 16
      pattern: "^.*?$"
    UserId:
      description: Identificatore unico dell'utente
      type: integer
      example: 1
      readOnly: true
...
    Text:
      type: string
      description: Contenuto testuale del messaggio
      example: "Ciao, come stai?"
      minLength: 1
      maxLength: 500
      pattern: '^.*?$'
  parameters:
    userId:
      schema:
        $ref: '#/components/schemas/UserId'
      name: userId
      in: path
      required: true
      description: ID utente
    conversationId:
      schema:
        type: integer
        example: 1
        readOnly: true
        description: "Identificatore unico della conversazione."
      name: conversationId
      in: path
      required: true
      description: ID conversazione
    messageId:
      schema:
        type: integer
        example: 1
        readOnly: true
        description: "Identificatore unico del messaggio."
      name: messageId
      in: path
      required: true
      description: ID messaggio
...
    groupId:
      schema:
        type: integer
        example: 1
        readOnly: true
        description: "Identificatore unico del gruppo."
      name: groupId
      in: path
      required: true
      description: ID gruppo
    sort:
      schema:
        type: string
        enum: [asc, desc]
        description: "Definisce l'ordine di sort dei messaggi."
      name: sort
      in: query
      required: true
      description: "Ordina per timestamp dell'ultimo messaggio. Usa 'desc' per ordine cronologico inverso, o 'asc' per ordine cronologico."
...